refactor(extraction): split extraction_service.py into src/services/extraction_service/ package
